Refactored code, added zoom factor for aux, improved bracketing in unlimited mode, changed sharpening algorithm to bilateral

+++ app/src/main/res/raw/averageff.glsl
@@ -1,6 +1,6 @@
 #version 300 es
 precision highp float;
-precision mediump sampler2D;
+precision highp sampler2D;
 uniform sampler2D InputBuffer;
 uniform sampler2D InputBuffer2;
 uniform int unlimitedcount;
